% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miprd-package.R
\docType{package}
\name{miprd-package}
\alias{miprd}
\alias{miprd-package}
\title{miprd: Construcción de Matrices Insumo-Producto desde Cuadros de Oferta y Utilización}
\description{
Convierte los Cuadros de Oferta y Utilización (COU) en Matrices Insumo-Producto (MIP) mediante el Modelo D y el Modelo B. Genera un archivo Excel con todas las matrices intermedias y finales etiquetadas.
}
\author{
\strong{Maintainer}: Briam Eduardo Guerrero Batista \email{briam.guerrero@bancentral.gov.do}

}
\keyword{internal}
